+++ benchmark/compute/dart_benchmark_rpc.dart
@@ -16,11 +16,11 @@ class IsolateRpcComputeBenchmark extends AsyncBenchmarkBase with BenchmarkTask, 
   IsolateRpcComputeBenchmark({required this.numOfTasks, required this.concurrency})
       : super('IsolateRpc,Compute(n=$numOfTasks,c=$concurrency)');
 
-  late RpcService<void, bool> rpcService;
+  late IsolateRpc<void, bool> rpc;
 
   @override
   Future<void> setup() {
-    rpcService = IsolateRpc.single(processor: (_) async {
+    rpc = IsolateRpc.single(processor: (_) async {
       return compute();
     });
     return Future.value();
@@ -28,7 +28,7 @@ class IsolateRpcComputeBenchmark extends AsyncBenchmarkBase with BenchmarkTask, 
 
   @override
   Future<void> teardown() async {
-    return rpcService.shutdown();
+    return rpc.shutdown();
   }
 
   @override
@@ -36,6 +36,6 @@ class IsolateRpcComputeBenchmark extends AsyncBenchmarkBase with BenchmarkTask, 
 
   @override
   Future<void> run() async {
-    return runTasks((i) => rpcService.execute(null));
+    return runTasks((i) => rpc.execute(null));
   }
 }