+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.0
+
+breaking change to refactor simpler API naming
+
+- Rename `RpcService` to `IsolateRpc`
+- Rename `RpcResponse` to `IsolateRpcResponse`
+- Rename `RpcError` to `IsolateRpcError`
+- Hide `IsolateRpc` constructor
+
 ## 1.0.1
 
 - Add more advance use case in README.md